LayoutTests/http/tests/site-isolation/page-lifecycle/pageswap-iframe.html

<!-- webkit-test-runner [ SiteIsolationEnabled=true ] -->

<!DOCTYPE html>
<title>Tests that pageswap is fired in iframe when site isolation is enabled</title>

<script src="/js-test-resources/js-test.js"></script>
<script>
description("Tests that pageswap is fired in iframe when site isolation is enabled.");
jsTestIsAsync = true;

const chan = new BroadcastChannel("channel");
chan.addEventListener("message", (e) => {
  caughtEvent = e.data;
  shouldBeEqualToString("caughtEvent.name", "pageswap");
  shouldBe("caughtEvent.params.activation", "null");
  shouldBe("caughtEvent.params.viewTransition", "null");
  finishJSTest();
})

</script>

<iframe src="/site-isolation/page-lifecycle/resources/navigating-away-subpage.html?event=pageswap&param=activation&param=viewTransition"></iframe>
